Teach foolscap to use twisted endpoints
Allow connection hints from the future!
No more filtering connection hints.
We use clientFromString to get the endpoint object
from the appropriate registered plugin.
We are also now managing connections with a connect deferred instead of an
IConnector... therefore we call the cancel method instead of disconnect.

This is work in progress; a few of the unit tests fail.