Remove host header from Negotiation
Reference: foolscap trac ticket #203
As part of an effort to port foolscap client
to twisted endpoints we need to not parse the
endpoint descriptor but instead pass it to clientFromString
and receive an endpoint object with which to handle
the transport. Now in TubConnector's connectToAll we no
longer pass the hostname into the TubConnectorClientFactory
constructor.